# Conflicts:
#	dimos/perception/detection/type/detection2d/__init__.py
#	dimos/perception/detection/type/detection2d/imageDetections2D.py


Former-commit-id: e94404d5bd8799631e7998697e2819f814092ee8
Former-commit-id: eeded3fe7ccde2f1f7eb108412cfbefe0140ffee

+++ dimos/perception/detection/person_tracker.py
@@ -40,7 +40,7 @@ class PersonTracker(Module):
 
     def center_to_3d(
         self,
-        pixel: tuple[int, int],
+        pixel: tuple[float, float],
         camera_info: CameraInfo,
         assumed_depth: float = 1.0,
     ) -> Vector3:
